scraping/detector-de-vagas/requirements.txt
kafka-python==2.0.2
psycopg2==2.9.6
requests==2.31